Major rewrite to the way shopkeepers get created and their data gets saved:
Changes to shopkeeper ids:
* 'Session id' is now named only 'id' , and persists across server
restarts / plugin reloads.
* Renamed getShopkeeper-methods to ShopkeeperRegistry#getShopkeeperById
and ShopkeeperRegistry#getShopkeeperByUniqueId
* Some minor refactoring to the way shopkeepers are created. The shop
types are no longer responsible for registering the created shopkeepers
(this happens inside the ShopkeeperRegistry now).

Changes to the loading and saving of shopkeepers:
* All shopkeeper data gets kept in memory now and during saves, only the
data of dirty shopkeepers gets updated.
  * This should come with a large performance improvement if there are
many shopkeepers loaded, but only few of them change between saves.
  * It is in the responsibility of the shopkeeper and the shop object
implementation to mark shopkeepers as dirty when appropriate.
  * Added Shopkeeper#save() and Shopkeeper#saveDelayed() as shortcuts
for marking a specific shopkeeper dirty and requesting a save of the
ShopkeeperStorage.
  * Saving debug messages shows number of dirty shopkeepers that were
saved.
* When legacy shopkeeper or shop object data gets found and imported,
the shopkeeper gets marked as dirty.
* During plugin enable a delayed save task gets started, that writes
back any dirty shopkeeper data (ex. imported legacy data).
* Renamed 'saveReal()' to 'saveNow()'.
* Added saveImmediate() to request a sync (blocking) save.
* When requesting a sync save (ex. during plugin disable), it always
waits now for async saves to finish, before checking and preparing the
sync save. This simplifies the saving (less concurrency), and also
allows reusing the SaveResult object.
* Saving aborts any delayed saving task now.
* No longer shutting the plugin down if a shopkeeper cannot be loaded.
The invalid shopkeeper data should get saved back to file again now,
without any information getting lost.
* No longer using default object type if a shopkeeper with invalid
object type is loaded. Instead the shopkeeper gets skipped during
loading.
* No longer defaulting player shops of unknown type to 'normal'. Instead
a warning gets logged and the shopkeeper gets skipped during loading.

Improvements to the life cycle handling of shopkeepers:
* The naming and calling of methods and events should not be clearer and
more consistent.
* Shopkeepers get properly unloaded now (with removal events) during
shutdown and reloads.
* Added createShopkeeper and loadShopkeeper to ShopkeeperRegistry API.
* Renamed ShopkeepersPlugin#createShopkeeper to
ShopkeepersPlugin#handleShopkeeperCreation to make it more clear that
this creates the shopkeeper with player limitations taken into account
and the creator receiving messages about a failed shopkeeper creation.
* Internal: Moved everything related to shopkeeper creation by players
into shop type.

Various changes to all shopkeeper events:
* Various shopkeeper events got slightly renamed.
* Added base event class for all events involving a shopkeeper.
* Added PlayerCreateShopkeeperEvent that gets called every time a player
creates a shopkeeper (including admin shopkeepers).
* Added ShopkeeperAddedEvent that gets called for the creation and
loading of all shopkeepers.
* Added ShopkeeperRemoveEvent that gets called for the deletion and
unloading of all shopkeepers.
* OpenTradeEvent was removed and replaced with a more generic
ShopkeeperOpenUIEvent

Other changes:
* Moved getMaxShops(Player) into Settings class
* Removed Shopkeeper#setLocation from API.
* Various refactoring.

+++ src/main/java/com/nisovin/shopkeepers/ui/defaults/EditorHandler.java
@@ -14,13 +14,14 @@ import com.nisovin.shopkeepers.AbstractShopkeeper;
 import com.nisovin.shopkeepers.Settings;
 import com.nisovin.shopkeepers.api.Shopkeeper;
 import com.nisovin.shopkeepers.api.ShopkeepersPlugin;
-import com.nisovin.shopkeepers.api.events.ShopkeeperDeletedEvent;
+import com.nisovin.shopkeepers.api.events.PlayerDeleteShopkeeperEvent;
 import com.nisovin.shopkeepers.api.events.ShopkeeperEditedEvent;
 import com.nisovin.shopkeepers.api.shopobjects.DefaultShopObjectTypes;
 import com.nisovin.shopkeepers.api.shoptypes.PlayerShopType;
 import com.nisovin.shopkeepers.ui.AbstractUIType;
 import com.nisovin.shopkeepers.ui.UIHandler;
 import com.nisovin.shopkeepers.util.ItemUtils;
+import com.nisovin.shopkeepers.util.Log;
 import com.nisovin.shopkeepers.util.Utils;
 
 public abstract class EditorHandler extends UIHandler {
@@ -48,8 +49,9 @@ public abstract class EditorHandler extends UIHandler {
 	@Override
 	protected void onInventoryClose(InventoryCloseEvent event, Player player) {
 		this.saveEditor(event.getInventory(), player);
-		this.getShopkeeper().closeAllOpenWindows();
-		ShopkeepersPlugin.getInstance().getShopkeeperStorage().save();
+		Shopkeeper shopkeeper = this.getShopkeeper();
+		shopkeeper.closeAllOpenWindows();
+		shopkeeper.save();
 	}
 
 	@Override
@@ -63,24 +65,28 @@ public abstract class EditorHandler extends UIHandler {
 			// delete button - delete shopkeeper:
 			event.setCancelled(true);
 
-			// return shop creation item for player shopkeepers:
-			if (Settings.deletingPlayerShopReturnsCreationItem && shopkeeper.getType() instanceof PlayerShopType) {
-				ItemStack shopCreationItem = Settings.createShopCreationItem();
-				Map<Integer, ItemStack> remaining = player.getInventory().addItem(shopCreationItem);
-				if (!remaining.isEmpty()) {
-					player.getWorld().dropItem(shopkeeper.getActualLocation(), shopCreationItem);
+			// call event:
+			PlayerDeleteShopkeeperEvent deleteEvent = new PlayerDeleteShopkeeperEvent(shopkeeper, player);
+			Bukkit.getPluginManager().callEvent(deleteEvent);
+			if (deleteEvent.isCancelled()) {
+				Log.debug("ShopkeeperDeleteEvent was cancelled!");
+			} else {
+				// return shop creation item for player shopkeepers:
+				if (Settings.deletingPlayerShopReturnsCreationItem && shopkeeper.getType() instanceof PlayerShopType) {
+					ItemStack shopCreationItem = Settings.createShopCreationItem();
+					Map<Integer, ItemStack> remaining = player.getInventory().addItem(shopCreationItem);
+					if (!remaining.isEmpty()) {
+						player.getWorld().dropItem(shopkeeper.getObjectLocation(), shopCreationItem);
+					}
 				}
+
+				// delete shopkeeper:
+				// this also deactivates the ui and closes all open windows for this shopkeeper after a delay
+				shopkeeper.delete();
+
+				// save:
+				shopkeeper.save();
 			}
-
-			// delete shopkeeper:
-			// this also deactivates the ui and closes all open windows for this shopkeeper after a delay
-			shopkeeper.delete();
-
-			// run event:
-			Bukkit.getPluginManager().callEvent(new ShopkeeperDeletedEvent(player, shopkeeper));
-
-			// save:
-			ShopkeepersPlugin.getInstance().getShopkeeperStorage().save();
 		} else if (rawSlot == 17) {
 			// cycle button - cycle to next object type variation:
 			event.setCancelled(true);
@@ -88,7 +94,7 @@ public abstract class EditorHandler extends UIHandler {
 			ItemStack cursor = event.getCursor();
 			if (!ItemUtils.isEmpty(cursor)) {
 				// equip item:
-				shopkeeper.getShopObject().setItem(cursor.clone());
+				shopkeeper.getShopObject().equipItem(cursor.clone());
 				// TODO how to remove equipped item again?
 				// TODO equipped items don't get saved current -> they get lost when the entity is respawned
 				// TODO not possible for player shops currently, because clicking/picking up items in player inventory
@@ -104,11 +110,11 @@ public abstract class EditorHandler extends UIHandler {
 				}
 			}
 
-			// run event:
-			Bukkit.getPluginManager().callEvent(new ShopkeeperEditedEvent(player, shopkeeper));
+			// call event:
+			Bukkit.getPluginManager().callEvent(new ShopkeeperEditedEvent(shopkeeper, player));
 
 			// save:
-			ShopkeepersPlugin.getInstance().getShopkeeperStorage().save();
+			shopkeeper.save();
 		} else if (rawSlot == 8) {
 			// naming or chest inventory button:
 			event.setCancelled(true);
@@ -126,11 +132,11 @@ public abstract class EditorHandler extends UIHandler {
 			// prepare closing the editor window:
 			this.saveEditor(event.getInventory(), player);
 
-			// run event:
-			Bukkit.getPluginManager().callEvent(new ShopkeeperEditedEvent(player, shopkeeper));
+			// call event:
+			Bukkit.getPluginManager().callEvent(new ShopkeeperEditedEvent(shopkeeper, player));
 
 			// save:
-			ShopkeepersPlugin.getInstance().getShopkeeperStorage().save();
+			shopkeeper.save();
 
 			// ignore other click events for this shopkeeper in the same tick:
 			shopkeeper.deactivateUI();
